third_party/blink/web_tests/external/wpt/css/cssom-view/scrollWidthHeightWhenNotScrollable-expected.txt

This is a testharness.js-based test.
[FAIL] elemOverflow.scrollHeight is the height of its scrolled contents (plus padding, since we overflowed)
  assert_equals: expected 200 but got 150
[FAIL] elemOverflow.scrollHeight is the width of its scrolled contents (plus padding)
  assert_equals: expected 290 but got 250
Harness: the test ran to completion.
